+++ src/game/tests/countering_and_mana.rs
@@ -1,4 +1,5 @@
 mod cavern_of_souls;
+mod protecting_lands;
 
 use super::*;
 use crate::AbilityProgramDef;